feat: allow 'graph' to be generated without 'target'
This switches off dependency resolution mode and dumps all the stages
(targets) as a graph.

While working on this feature, I discovered two cases when parts of the
build graph were duplicated: first when packages were resolved, leading
to duplicate nodes in the graph, second when LLB is generated. Buildkit
seems to handle correctly LLB de-duplication, but anyway it's better to
skip duplicate entries.

Fixes #22

As examples, see talos-systems/tools#73 and talos-systems/pkgs#46

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>