feat: add support for runtime dependencies
Fixes #30

This adds `runtime: <bool>` flag to all the dependencies. Runtime
dependencies are implicitly in any stages using this stage (including
transitive runtime dependencies). This allows to dry up dependencies as
we list them in `pkg.yaml`.

Dependency code was rewritten while implementing this change, now
internal and external dependencies are handled in the same list.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>