feat: stop on error in any package, more validation
Pack of small fixes to `bldr` discovered while working on the
tests.

When loading package fails, it gets logged, but when using frontend mode
this log is hard to reach: it goes into buildkitd docker container log.
Skipped package would probably result in some dependency not being
resolvable, and actual error would be buried deep in the log.

Instead of that, abort processing and return an error which is printed
to console log when processing dockerfile frontend.

Do more validation for `pkg.yaml` as invalid field values lead to
cryptic LLB errors which are hard to match with `pkg.yaml` problems.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>