fix: support multi-platform key passed in via options
When multi-arch build is split across builders, each builder might get a
list with only a single platform, but the fact that overall build is
mulit-platform is signalled via a key. If key is set, frontend acts as
if it builds multi-platform image: export maps are generated, proper
prefixes are added to the build steps.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>