feat: optimize build time by removing unnecessary I/O
This PR contains two optimizations:

* reorder steps to make variant + dependencies layers cacheable
* promote first dependency to base image for 'scratch' builds

First optimization is pretty simple, given that many packages contain
same top dependencies:

```yaml
dependencies:
  - stage: base
```

It makes sure LLB is ordered so that base part of the LLB up to the
dependencies is identical across packages builds and can be eliminated
to a single instance.

Second optimization is more tricky, for `scratch` packages:

```yaml
variant: scratch
dependencies:
  - stage: tools
  - stage: libressl
```

It promotes `tools` to be base image (instead of scratch). To compare,
in `Dockerfile` syntax it means replacing:

```
FROM scratch
COPY --from=tools / /
COPY --from=libressl / /
```

with:

```
FROM tools
COPY --from=libressl / /
```

If the `tools` layer is huge (which is often the case, in order of 0.5
GiB), it removes a lot of file copy operations and improves build time.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>