feat: implement `SOURCE_DATE_EPOCH` build argument
In general, buildkit ignores timestamps when caching intermediate build
results, but still timestamps leak into the final container image.

So if the package is rebuilt with cache, output image stays same (even
though build step might produce different timestamps, but same content,
final output will still be cached).

This new feature enforces timestamp from `SOURCE_DATE_EPOCH` on the
output in the `finalize` step so that build is reproducible.

The downside is that if the build step is used in other steps, and
`SOURCE_DATE_EPOCH` is used, it might cause cascading rebuilds.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>