feat: implement additional variables and `bldr eval`
The goal of this PR is to enable custom image tags for extensions.

This PR adds deep nested templated additional variables per package
(more precisely, per directory) `vars.yaml` which might be used to
template `pkg.yaml`.

New command `bldr eval` can be used to output value of any variable.

Build arguments are injected as `BUILD_ARG_` variables.

The idea is to declare extension version as a variable, and use that
variable to produce image tag and in `pkg.yaml`.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>