fix(download): always send User-Agent and Accept headers
When downloading files, src.fedoraproject.org now requires the Accept
header to be present. If the header is missing, it returns an HTML page
instead of the requested file, causing a checksum mismatch.

This change adds the "Accept: */*" and "User-Agent: BLDR/<version>"
headers to all HTTP downloads to prevent this issue.

This was discovered when glibc and libtirpc packages started to fail
recently in the extensions repository[1].

[1] https://github.com/siderolabs/extensions/

Signed-off-by: Gilbert Gilb's <gilbsgilbert@gmail.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>